Extract shared Playwright route readiness helpers
